Refactor path to brand-utils.js
