+++ src/components/CookieWarning.js
@@ -2,9 +2,9 @@ import _ from 'lodash/fp'
 import { useEffect, useRef, useState } from 'react'
 import { aside, div, h } from 'react-hyperscript-helpers'
 import { Transition } from 'react-transition-group'
+import { getEnabledBrand } from 'src/brand-utils'
 import { ButtonPrimary, ButtonSecondary, Link } from 'src/components/common'
 import { Ajax } from 'src/libs/ajax'
-import { getEnabledBrand } from 'src/libs/brand-utils'
 import { getSessionStorage } from 'src/libs/browser-storage'
 import colors from 'src/libs/colors'
 import * as Nav from 'src/libs/nav'
