+++ src/components/Favicon.js
@@ -1,7 +1,7 @@
 import RFavicon from 'react-favicon'
 import { h } from 'react-hyperscript-helpers'
+import { isBioDataCatalyst } from 'src/brand-utils'
 import bioDataCatalystFavicon from 'src/images/brands/bioDataCatalyst/bioDataCatalyst-favicon.jpg'
-import { isBioDataCatalyst } from 'src/libs/brand-utils'
 import * as Utils from 'src/libs/utils'
