+++ src/components/FooterWrapper.js
@@ -1,8 +1,8 @@
 import { Fragment, useEffect, useState } from 'react'
 import { a, div, h } from 'react-hyperscript-helpers'
+import { isBioDataCatalyst, isRadX } from 'src/brand-utils'
 import { Clickable, Link } from 'src/components/common'
 import { icon } from 'src/components/icons'
-import { isBioDataCatalyst, isRadX } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
 import { footerLogo } from 'src/libs/logos'
 import * as Nav from 'src/libs/nav'
