+++ src/components/NpsSurvey.js
@@ -2,12 +2,12 @@ import { differenceInCalendarDays, parseJSON } from 'date-fns/fp'
 import _ from 'lodash/fp'
 import { useEffect, useState } from 'react'
 import { div, h, input, span } from 'react-hyperscript-helpers'
+import { getEnabledBrand } from 'src/brand-utils'
 import { ButtonSecondary, Clickable } from 'src/components/common'
 import { icon } from 'src/components/icons'
 import { TextArea } from 'src/components/input'
 import Interactive from 'src/components/Interactive'
 import { Ajax } from 'src/libs/ajax'
-import { getEnabledBrand } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
 import { withErrorIgnoring } from 'src/libs/error'
 import { useStore } from 'src/libs/react-utils'
