+++ src/components/common.js
@@ -7,6 +7,7 @@ import { b, div, h, h1, img, input, label, span } from 'react-hyperscript-helper
 import RSelect, { components as RSelectComponents } from 'react-select'
 import RAsyncCreatableSelect from 'react-select/async-creatable'
 import RSwitch from 'react-switch'
+import { getEnabledBrand, isRadX } from 'src/brand-utils'
 import FooterWrapper from 'src/components/FooterWrapper'
 import { centeredSpinner, containsUnlabelledIcon, icon } from 'src/components/icons'
 import { TextInput } from 'src/components/input'
@@ -18,7 +19,6 @@ import TopBar from 'src/components/TopBar'
 import landingPageHero from 'src/images/landing-page-hero.jpg'
 import scienceBackground from 'src/images/science-background.jpg'
 import { Ajax } from 'src/libs/ajax'
-import { getEnabledBrand, isRadX } from 'src/libs/brand-utils'
 import colors, { terraSpecial } from 'src/libs/colors'
 import { getConfig } from 'src/libs/config'
 import { withErrorReporting } from 'src/libs/error'
