+++ src/libs/link-expiration-alerts.js
@@ -2,8 +2,8 @@ import { addDays, differenceInDays, parseJSON } from 'date-fns/fp'
 import _ from 'lodash/fp'
 import { Fragment, useEffect, useState } from 'react'
 import { h } from 'react-hyperscript-helpers'
+import { getEnabledBrand } from 'src/brand-utils'
 import { FrameworkServiceLink, ShibbolethLink, UnlinkFenceAccount } from 'src/components/common'
-import { getEnabledBrand } from 'src/libs/brand-utils'
 import * as Nav from 'src/libs/nav'
 import allProviders from 'src/libs/providers'
 import { authStore } from 'src/libs/state'
