+++ src/pages/LandingPage.js
@@ -1,11 +1,11 @@
 import _ from 'lodash/fp'
 import { div, h, h2 } from 'react-hyperscript-helpers'
+import { getEnabledBrand, isFirecloud, isTerra } from 'src/brand-utils'
 import { ButtonOutline, Clickable, HeroWrapper, Link } from 'src/components/common'
 import { icon } from 'src/components/icons'
 import hexButton from 'src/images/hex-button.svg'
 import terraHero from 'src/images/terra-hero.png'
 import { Ajax } from 'src/libs/ajax'
-import { getEnabledBrand, isFirecloud, isTerra } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
 import Events from 'src/libs/events'
 import * as Nav from 'src/libs/nav'
