+++ src/pages/SignIn.js
@@ -1,8 +1,8 @@
 import { Fragment } from 'react'
 import { div, h, p } from 'react-hyperscript-helpers'
+import { getEnabledBrand, isAnvil, isBioDataCatalyst, isElwazi, isFirecloud } from 'src/brand-utils'
 import { HeroWrapper, Link } from 'src/components/common'
 import SignInButton from 'src/components/SignInButton'
-import { getEnabledBrand, isAnvil, isBioDataCatalyst, isElwazi, isFirecloud } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
