+++ src/pages/library/Datasets.js
@@ -1,5 +1,6 @@
 import { Fragment, useState } from 'react'
 import { b, div, h, img, p, span } from 'react-hyperscript-helpers'
+import { getEnabledBrand, isRadX } from 'src/brand-utils'
 import { ButtonPrimary, Link } from 'src/components/common'
 import FooterWrapper from 'src/components/FooterWrapper'
 import { libraryTopMatter } from 'src/components/library-common'
@@ -21,7 +22,6 @@ import targetLogo from 'src/images/library/datasets/target_logo.jpeg'
 import tcgaLogo from 'src/images/library/datasets/TCGALogo.jpg'
 import topMedLogo from 'src/images/library/datasets/TopMed@2x.png'
 import { Ajax } from 'src/libs/ajax'
-import { getEnabledBrand, isRadX } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
 import { getConfig } from 'src/libs/config'
 import Events from 'src/libs/events'
