+++ src/pages/workspaces/workspace/Dashboard.js
@@ -1,6 +1,7 @@
 import _ from 'lodash/fp'
 import { Fragment, useCallback, useEffect, useImperativeHandle, useRef, useState } from 'react'
 import { dd, div, dl, dt, h, h3, i, span, strong } from 'react-hyperscript-helpers'
+import { getEnabledBrand } from 'src/brand-utils'
 import * as breadcrumbs from 'src/components/breadcrumbs'
 import { requesterPaysWrapper } from 'src/components/bucket-utils'
 import Collapse from 'src/components/Collapse'
@@ -19,7 +20,6 @@ import { ReactComponent as GcpLogo } from 'src/images/gcp.svg'
 import { Ajax } from 'src/libs/ajax'
 import { bucketBrowserUrl, refreshTerraProfile } from 'src/libs/auth'
 import { getRegionFlag, getRegionLabel } from 'src/libs/azure-utils'
-import { getEnabledBrand } from 'src/libs/brand-utils'
 import colors from 'src/libs/colors'
 import { reportError, withErrorReporting } from 'src/libs/error'
 import Events from 'src/libs/events'
